Re-write onTouchEvent logic for joysticks and spells
Eldered Jonas @ Youtube wrote a comment in one of my videos, asking
how one would go about when adding an extra joystick/thumbstick to
the game. He could visually add a new joystick, following my previous
joystick example, but he couldn't get the additional joystick to
respond on touch input. To solved this, we had some design converstaions
and finally, we had a working onTouchEvent implementation what supported
additional joysticks, by keeping track of pointer IDs etc.

The new directional joystick added in this commit is not yet in control
of the players diretion. It was chosen to wait with that part to make
this commit smaller, and to have a clearer division of the respective
work needed for the complete feature, since it would become a very long
YouTube video if I covered the whole implementation in one session.