Add GameDisplay
The GameDisplay is a class which enables for drawing the game world centered around an object, such as the player. The reason of centering the world around the player is that it's now possible to move around on an infinitely large map without having the player dissapear outside of the screen.